move utils to tests/api module
